Moved I2PEndian to util/
