+++ Garlic.cpp
@@ -1,5 +1,5 @@
 #include <inttypes.h>
-#include "I2PEndian.h"
+#include "util/I2PEndian.h"
 #include <map>
 #include <string>
 #include "RouterContext.h"
