+++ HTTPProxy.cpp
@@ -10,7 +10,7 @@
 #include "Streaming.h"
 #include "Destination.h"
 #include "ClientContext.h"
-#include "I2PEndian.h"
+#include "util/I2PEndian.h"
 #include "I2PTunnel.h"
 
 namespace i2p
