+++ Reseed.cpp
@@ -11,7 +11,7 @@
 #include <cryptopp/zinflate.h>
 #define CRYPTOPP_ENABLE_NAMESPACE_WEAK 1
 #include <cryptopp/arc4.h>
-#include "I2PEndian.h"
+#include "util/I2PEndian.h"
 #include "Reseed.h"
 #include "Log.h"
 #include "Identity.h"
