+++ TunnelEndpoint.cpp
@@ -1,4 +1,4 @@
-#include "I2PEndian.h"
+#include "util/I2PEndian.h"
 #include <string.h>
 #include "Log.h"
 #include "NetDb.h"
