+++ util/I2PEndian.cpp
